+++ hooks/index.ts
@@ -1,5 +1,6 @@
 import Accordion from "./accordion";
 import Clipboard from "./clipboard";
+import Collapsible from "./collapsible";
 import Combobox from "./combobox";
 import Dialog from "./dialog";
 import Menu from "./menu";
@@ -8,10 +9,11 @@ import PinInput from "./pin-input";
 const Hooks = {
   Accordion,
   Clipboard,
+  Collapsible,
   Combobox,
   Dialog,
   Menu,
   PinInput,
 };
 
-export { Accordion, Clipboard, Combobox, Dialog, Menu, PinInput, Hooks };
+export { Accordion, Clipboard, Collapsible, Combobox, Dialog, Menu, PinInput, Hooks };